requirements.txt
av==14.4.0
dataclasses_json==0.6.7
decord==0.6.0
Flask==3.1.1
flask_cors==6.0.1
huggingface_hub==0.27.1
hydra-core==1.3.2
imagesize==1.4.1
iopath==0.1.10
matplotlib==3.10.3
mmcv-full
numpy==2.3.1
omegaconf==2.3.0
opencv_python==4.10.0.84
pandas==2.3.0
Pillow==11.3.0
pycocotools==2.0.8
scipy==1.16.0
setuptools==75.1.0
scikit-image
strawberry==3.0
submitit==1.5.3
tensordict==0.6.2
torch
torchvision
tqdm==4.67.1
